internal/json: rename internal/contour to internal/json
Rename internal/contour (which only contains the JSON xDS API) to
internal/json which more accurately reflects its contents.
