+++ internal/json/sds.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package contour
+package json
 
 import (
 	"github.com/heptio/contour/internal/envoy"
